consumer/requirements.txt
sqlalchemy
psycopg2
kafka-python
pytest